Add tests for third party table with long identifiers
