+++ mathesar/tests/api/test_column_api.py
@@ -9,9 +9,6 @@ from db.types.base import PostgresType, MathesarCustomType
 
 from mathesar.api.exceptions.error_codes import ErrorCodes
 from mathesar.tests.api.test_table_api import check_columns_response
-from mathesar.api.exceptions.database_exceptions import (
-    exceptions as database_api_exceptions
-)
 
 
 def test_column_list(column_test_table, client):
@@ -171,22 +168,6 @@ def test_column_create(table_fixture, client, request):
     assert actual_new_col["name"] == name
     assert actual_new_col["type"] == db_type.id
     assert actual_new_col["default"] is None
-
-
-def test_column_create_with_long_column_name(column_test_table, client):
-    very_long_string = ''.join(map(str, range(50)))
-    name = 'very_long_identifier_' + very_long_string
-    db_type = PostgresType.NUMERIC
-    data = {
-        "name": name,
-        "type": db_type.id,
-    }
-    response = client.post(
-        f"/api/db/v0/tables/{column_test_table.id}/columns/",
-        data=data,
-    )
-    assert response.status_code == 400
-    assert response.json()[0]['code'] == database_api_exceptions.IdentifierTooLong.error_code
 
 
 @pytest.mark.parametrize('client_name, expected_status_code', write_client_with_different_roles)
